+++ sketch-core/src/iosTest/kotlin/com/github/panpf/sketch/core/ios/test/fetch/PhotosAssetUriFetcherTest.kt
@@ -1,23 +1,31 @@
 package com.github.panpf.sketch.core.ios.test.fetch
 
+import com.github.panpf.sketch.cache.CachePolicy
+import com.github.panpf.sketch.cache.DiskCache
 import com.github.panpf.sketch.fetch.PhotosAssetUriFetcher
 import com.github.panpf.sketch.fetch.isPhotosAssetUri
 import com.github.panpf.sketch.fetch.newPhotosAssetUri
 import com.github.panpf.sketch.fetch.parseLocalIdentifier
 import com.github.panpf.sketch.request.ImageRequest
 import com.github.panpf.sketch.request.allowNetworkAccessPhotosAsset
+import com.github.panpf.sketch.request.preferFileCacheForImagePhotosAsset
 import com.github.panpf.sketch.request.preferThumbnailForPhotosAsset
+import com.github.panpf.sketch.request.useSkiaForImagePhotosAsset
+import com.github.panpf.sketch.test.singleton.getSketch
 import com.github.panpf.sketch.test.singleton.getTestContextAndSketch
+import com.github.panpf.sketch.test.utils.getTestContext
 import com.github.panpf.sketch.test.utils.toRequestContext
 import com.github.panpf.sketch.util.Size
 import com.github.panpf.sketch.util.toUri
 import kotlinx.coroutines.test.runTest
+import okio.FileSystem
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
+import kotlin.test.assertSame
 import kotlin.test.assertTrue
 
 class PhotosAssetUriFetcherTest {
@@ -57,12 +65,25 @@ class PhotosAssetUriFetcherTest {
 
     @Test
     fun testConstructor() {
+        val sketch = getSketch()
         val localIdentifier = "DB16113B-984A-4D12-B4D0-50FC46066781/L0/001"
-        PhotosAssetUriFetcher(localIdentifier, false, true)
+        PhotosAssetUriFetcher(
+            localIdentifier,
+            false,
+            true,
+            true,
+            false,
+            sketch.downloadCache,
+            CachePolicy.WRITE_ONLY,
+        )
         PhotosAssetUriFetcher(
             localIdentifier = localIdentifier,
             preferredThumbnail = false,
-            allowNetworkAccess = true
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = sketch.downloadCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY,
         )
     }
 
@@ -75,46 +96,194 @@ class PhotosAssetUriFetcherTest {
 
     @Test
     fun testFetch() = runTest {
+        val sketch = getSketch()
         val localIdentifier = "DB16113B-984A-4D12-B4D0-50FC46066781/L0/001"
         val result = PhotosAssetUriFetcher(
             localIdentifier = localIdentifier,
             preferredThumbnail = false,
-            allowNetworkAccess = true
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = sketch.downloadCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY,
         ).fetch()
         assertFalse(result.isSuccess)
     }
 
     @Test
     fun testEqualsAndHashCode() {
+        val context = getTestContext()
         val localIdentifier = "DB16113B-984A-4D12-B4D0-50FC46066781/L0/001"
         val localIdentifier2 = "DB16113B-984A-4D12-B4D0-50FC46066781/L0/002"
-        val factory1 = PhotosAssetUriFetcher(localIdentifier, false, true)
-        val factory11 = PhotosAssetUriFetcher(localIdentifier, false, true)
-        val factory2 = PhotosAssetUriFetcher(localIdentifier2, false, true)
-        val factory3 = PhotosAssetUriFetcher(localIdentifier, true, true)
-        val factory4 = PhotosAssetUriFetcher(localIdentifier, false, false)
+        val diskCache = DiskCache.Builder(context, FileSystem.SYSTEM, "test_cache").build()
+        val diskCache2 = DiskCache.Builder(context, FileSystem.SYSTEM, "test_cache2").build()
+        val factory1 = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier,
+            preferredThumbnail = false,
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = diskCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY
+        )
+        val factory11 = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier,
+            preferredThumbnail = false,
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = diskCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY
+        )
+        val factory2 = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier2,
+            preferredThumbnail = false,
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = diskCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY
+        )
+        val factory3 = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier,
+            preferredThumbnail = true,
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = diskCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY
+        )
+        val factory4 = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier,
+            preferredThumbnail = false,
+            allowNetworkAccess = false,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = diskCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY
+        )
+        val factory5 = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier,
+            preferredThumbnail = false,
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = false,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = diskCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY
+        )
+        val factory6 = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier,
+            preferredThumbnail = false,
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = true,
+            downloadCache = diskCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY
+        )
+        val factory7 = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier,
+            preferredThumbnail = false,
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = diskCache2,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY
+        )
+        val factory8 = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier,
+            preferredThumbnail = false,
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = diskCache,
+            downloadCachePolicy = CachePolicy.READ_ONLY
+        )
 
         assertEquals(expected = factory1, actual = factory11)
         assertNotEquals(illegal = factory1, actual = factory2)
+        assertNotEquals(illegal = factory1, actual = factory3)
+        assertNotEquals(illegal = factory1, actual = factory4)
+        assertNotEquals(illegal = factory1, actual = factory5)
+        assertNotEquals(illegal = factory1, actual = factory6)
+        assertNotEquals(illegal = factory1, actual = factory7)
+        assertNotEquals(illegal = factory1, actual = factory8)
         assertNotEquals(illegal = factory2, actual = factory3)
         assertNotEquals(illegal = factory2, actual = factory4)
+        assertNotEquals(illegal = factory2, actual = factory5)
+        assertNotEquals(illegal = factory2, actual = factory6)
+        assertNotEquals(illegal = factory2, actual = factory7)
+        assertNotEquals(illegal = factory2, actual = factory8)
         assertNotEquals(illegal = factory3, actual = factory4)
+        assertNotEquals(illegal = factory3, actual = factory5)
+        assertNotEquals(illegal = factory3, actual = factory6)
+        assertNotEquals(illegal = factory3, actual = factory7)
+        assertNotEquals(illegal = factory3, actual = factory8)
+        assertNotEquals(illegal = factory4, actual = factory5)
+        assertNotEquals(illegal = factory4, actual = factory6)
+        assertNotEquals(illegal = factory4, actual = factory7)
+        assertNotEquals(illegal = factory4, actual = factory8)
+        assertNotEquals(illegal = factory5, actual = factory6)
+        assertNotEquals(illegal = factory5, actual = factory7)
+        assertNotEquals(illegal = factory5, actual = factory8)
+        assertNotEquals(illegal = factory6, actual = factory7)
+        assertNotEquals(illegal = factory6, actual = factory8)
+        assertNotEquals(illegal = factory7, actual = factory8)
         assertNotEquals(illegal = factory1, actual = null as Any?)
         assertNotEquals(illegal = factory1, actual = Any())
 
         assertEquals(expected = factory1.hashCode(), actual = factory11.hashCode())
         assertNotEquals(illegal = factory1.hashCode(), actual = factory2.hashCode())
+        assertNotEquals(illegal = factory1.hashCode(), actual = factory3.hashCode())
+        assertNotEquals(illegal = factory1.hashCode(), actual = factory4.hashCode())
+        assertNotEquals(illegal = factory1.hashCode(), actual = factory5.hashCode())
+        assertNotEquals(illegal = factory1.hashCode(), actual = factory6.hashCode())
+        assertNotEquals(illegal = factory1.hashCode(), actual = factory7.hashCode())
+        assertNotEquals(illegal = factory1.hashCode(), actual = factory8.hashCode())
         assertNotEquals(illegal = factory2.hashCode(), actual = factory3.hashCode())
         assertNotEquals(illegal = factory2.hashCode(), actual = factory4.hashCode())
+        assertNotEquals(illegal = factory2.hashCode(), actual = factory5.hashCode())
+        assertNotEquals(illegal = factory2.hashCode(), actual = factory6.hashCode())
+        assertNotEquals(illegal = factory2.hashCode(), actual = factory7.hashCode())
+        assertNotEquals(illegal = factory2.hashCode(), actual = factory8.hashCode())
         assertNotEquals(illegal = factory3.hashCode(), actual = factory4.hashCode())
+        assertNotEquals(illegal = factory3.hashCode(), actual = factory5.hashCode())
+        assertNotEquals(illegal = factory3.hashCode(), actual = factory6.hashCode())
+        assertNotEquals(illegal = factory3.hashCode(), actual = factory7.hashCode())
+        assertNotEquals(illegal = factory3.hashCode(), actual = factory8.hashCode())
+        assertNotEquals(illegal = factory4.hashCode(), actual = factory5.hashCode())
+        assertNotEquals(illegal = factory4.hashCode(), actual = factory6.hashCode())
+        assertNotEquals(illegal = factory4.hashCode(), actual = factory7.hashCode())
+        assertNotEquals(illegal = factory4.hashCode(), actual = factory8.hashCode())
+        assertNotEquals(illegal = factory5.hashCode(), actual = factory6.hashCode())
+        assertNotEquals(illegal = factory5.hashCode(), actual = factory7.hashCode())
+        assertNotEquals(illegal = factory5.hashCode(), actual = factory8.hashCode())
+        assertNotEquals(illegal = factory6.hashCode(), actual = factory7.hashCode())
+        assertNotEquals(illegal = factory6.hashCode(), actual = factory8.hashCode())
+        assertNotEquals(illegal = factory7.hashCode(), actual = factory8.hashCode())
     }
 
     @Test
     fun testToString() {
+        val sketch = getSketch()
         val localIdentifier = "DB16113B-984A-4D12-B4D0-50FC46066781/L0/001"
-        val fetcher = PhotosAssetUriFetcher(localIdentifier, false, true)
+        val fetcher = PhotosAssetUriFetcher(
+            localIdentifier = localIdentifier,
+            preferredThumbnail = false,
+            allowNetworkAccess = true,
+            useSkiaForImagePhotosAsset = true,
+            preferFileCacheForImagePhotosAsset = false,
+            downloadCache = sketch.downloadCache,
+            downloadCachePolicy = CachePolicy.WRITE_ONLY
+        )
         assertEquals(
-            expected = "PhotosAssetUriFetcher(localIdentifier='$localIdentifier', preferredThumbnail=false, allowNetworkAccess=true)",
+            expected = "PhotosAssetUriFetcher(" +
+                    "localIdentifier='$localIdentifier', " +
+                    "preferredThumbnail=false, " +
+                    "allowNetworkAccess=true, " +
+                    "useSkiaForImagePhotosAsset=true, " +
+                    "preferFileCacheForImagePhotosAsset=false, " +
+                    "downloadCache=${sketch.downloadCache}, " +
+                    "downloadCachePolicy=WRITE_ONLY)",
             actual = fetcher.toString()
         )
     }
@@ -131,7 +300,7 @@ class PhotosAssetUriFetcherTest {
         val localIdentifier = "DB16113B-984A-4D12-B4D0-50FC46066781/L0/001"
         val imageUri = newPhotosAssetUri(localIdentifier)
 
-        ImageRequest(context, "htt://sample.com/sample.jpg")
+        ImageRequest(context, "http://sample.com/sample.jpg")
             .toRequestContext(sketch, Size.Empty)
             .let { factory.create(it) }
             .apply {
@@ -146,11 +315,18 @@ class PhotosAssetUriFetcherTest {
                 assertEquals(localIdentifier, this.localIdentifier)
                 assertFalse(this.preferredThumbnail)
                 assertFalse(this.allowNetworkAccess)
+                assertFalse(this.useSkiaForImagePhotosAsset)
+                assertFalse(this.preferFileCacheForImagePhotosAsset)
+                assertSame(sketch.downloadCache, this.downloadCache)
+                assertEquals(CachePolicy.ENABLED, this.downloadCachePolicy)
             }
 
         ImageRequest(context, imageUri) {
             preferThumbnailForPhotosAsset(true)
             allowNetworkAccessPhotosAsset(true)
+            useSkiaForImagePhotosAsset()
+            preferFileCacheForImagePhotosAsset()
+            downloadCachePolicy(CachePolicy.READ_ONLY)
         }.toRequestContext(sketch, Size.Empty)
             .let { factory.create(it) }
             .apply {
@@ -158,6 +334,10 @@ class PhotosAssetUriFetcherTest {
                 assertEquals(localIdentifier, this.localIdentifier)
                 assertTrue(this.preferredThumbnail)
                 assertTrue(this.allowNetworkAccess)
+                assertTrue(this.useSkiaForImagePhotosAsset)
+                assertTrue(this.preferFileCacheForImagePhotosAsset)
+                assertSame(sketch.downloadCache, this.downloadCache)
+                assertEquals(CachePolicy.READ_ONLY, this.downloadCachePolicy)
             }
     }